+++ src/background.js
@@ -1,4 +1,62 @@
-function createGotoMenu(scheme, url) {
+function createSearchMenuItem(msg, currentProvider, silent) {
+  browser.contextMenus.create({
+    id: currentProvider.name,
+    title: `${browser.i18n.getMessage('searchWith')} ${currentProvider.name}: ${msg}`,
+    contexts: ['selection', 'link', 'editable'],
+    onclick() {
+      const query = encodeURIComponent(msg);
+      browser.tabs.create({
+        url: `${currentProvider.url}${query}`,
+        active: !silent,
+      });
+    },
+  });
+}
+function updateSearchMenuItem(msg, currentProvider, silent) {
+  browser.contextMenus.update(currentProvider.name, {
+    title: `${browser.i18n.getMessage('searchWith')} ${currentProvider.name}: ${msg}`,
+    onclick() {
+      const query = encodeURIComponent(msg);
+      browser.tabs.create({
+        url: `${currentProvider.url}${query}`,
+        active: !silent,
+      });
+    },
+  });
+}
+
+function removeMenu(ids) {
+  ids.forEach((element) => {
+    if (element) browser.contextMenus.remove(element);
+  });
+}
+
+function refreshSearchMenu(msg, currentProvider, ids, silent) {
+  const menuItems = [];
+
+  if (!ids) {
+    ids = [];// eslint-disable-line
+  }
+  currentProvider.forEach((provider) => {
+    const index = ids.indexOf(provider.name);
+
+    menuItems.push(provider.name);
+    if (index !== -1) {
+      updateSearchMenuItem(msg, provider, silent);
+      ids[index] = undefined;// eslint-disable-line
+    } else {
+      createSearchMenuItem(msg, provider, silent);
+    }
+  });
+
+  removeMenu(ids);
+
+  return menuItems;
+}
+
+function createGotoMenu(scheme, url, settings) {
+  const opts = settings || {};
+
   browser.contextMenus.create({
     id: 'contextGoto',
     title: `${browser.i18n.getMessage('goTo')}: "${scheme}${url}"`,
@@ -6,17 +64,21 @@ function createGotoMenu(scheme, url) {
     onclick() {
       browser.tabs.create({
         url: `${scheme}${url}`,
+        active: !opts.silent,
       });
     },
   });
 }
-function updateGotoMenu(scheme, url) {
+function updateGotoMenu(scheme, url, settings) {
+  const opts = settings || {};
+
   browser.contextMenus.update('contextGoto', {
     title: `${browser.i18n.getMessage('goTo')}: "${scheme}${url}"`,
     contexts: ['selection', 'link', 'editable'],
     onclick() {
       browser.tabs.create({
         url: `${scheme}${url}`,
+        active: !opts.silent,
       });
     },
   });
@@ -37,10 +99,12 @@ function handleMessage(msg) {
   if (locals.lastMsg !== msg) {
     locals.lastMsg = msg;
     const {
-      defaultProtocol,
-      // TODO (!) add setting and new var urlDetected
-      urlDetected,
       currentProvider,
+      defaultProtocol,
+      gotoMenu,
+      mode,
+      searchMenu,
+      silent,
     } = locals;
 
     if (msg) {
@@ -49,98 +113,124 @@ function handleMessage(msg) {
       const urlWithHostnameRegexp = /^(?:(?:\S+(?::\S*)?@)?(?:(?:[a-z]+[a-z\d-]*(?:\.[a-z]+[a-z\d-]*)+)|(?:\d{1,3}(?:\.\d{1,3}){3}))(?::\d+)?)(?:(?:\/[^/\s#?]+)+\/?|\/)?(?:\?[^#\s]*)?(?:#[^\s]*)?$/gi;
       const msgForTest = msg.trim().toLowerCase();
 
-      browser.contextMenus.update('contextSearch', { // TODO Refactor move function outsides
-        title: `${browser.i18n.getMessage('searchWith')} ${currentProvider.name}: ${msg}`,
-        contexts: ['selection', 'link', 'editable'],
-        onclick() {
-          const query = encodeURIComponent(msg);
-          browser.tabs.create({
-            url: `${currentProvider.url}${query}`,
-          });
-        },
-      });
+      let urlDetected = false;
+      let protocol = defaultProtocol;
 
       if (urlWithSchemeRegexp.test(msgForTest)) {
-        if (urlDetected) {
-          updateGotoMenu('', msgForTest);
-        } else {
-          createGotoMenu('', msgForTest);
-          locals.urlDetected = true;
-        }
+        urlDetected = true;
+        protocol = '';
       } else if (urlWithHostnameRegexp.test(msgForTest)) {
-        if (urlDetected) {
-          updateGotoMenu(defaultProtocol, msgForTest);
+        urlDetected = true;
+      }
+
+      if (mode === '3' || (mode === '4' && !urlDetected) || mode === '1') {
+        locals.searchMenu = refreshSearchMenu(msg, currentProvider, searchMenu, silent);
+      } else if (searchMenu) {
+        removeMenu(searchMenu);
+        locals.searchMenu = false;
+      }
+
+      if (mode !== '1' && urlDetected) {
+        if (gotoMenu) {
+          updateGotoMenu(protocol, msgForTest, { silent });
         } else {
-          createGotoMenu(defaultProtocol, msgForTest);
-          locals.urlDetected = true;
+          createGotoMenu(protocol, msgForTest, { silent });
+          locals.gotoMenu = true;
         }
-      } else if (locals.urlDetected) {
+      } else if (gotoMenu) {
         browser.contextMenus.remove('contextGoto');
-        locals.urlDetected = false;
+        locals.gotoMenu = false;
+      }
+
+      if (locals.searchMenu && (locals.searchMenu.length > 1 || locals.gotoMenu)) {
+        currentProvider.forEach((element) => {
+          browser.contextMenus.update(element.name, {
+            title: `${element.name}: ${msg}`, // short title replace
+          });
+        });
       }
     } else {
-      browser.contextMenus.update('contextSearch', {
-        contexts: ['selection'],
-      });
-      if (urlDetected) {
-        browser.contextMenus.remove('contextGoto');
-        locals.urlDetected = false;
-      }
+      browser.contextMenus.removeAll();
+      locals.gotoMenu = false;
+      locals.searchMenu = false;
     }
   }
 }
 
 function onStorageChange(changes) {
   const locals = window.contextMenuSearchLocals;
-  const { lastMsg } = locals;
+  const { currentProvider } = changes;
 
-  locals.lastMsg += 'changed';
+  locals.lastMsg += 'changed'; // 'hack' to force update context menus (in handleMessage()) after settings changing
   // TODO think to remove eslint ignore comments
   delete changes.providers;// eslint-disable-line
 
-  if (changes.currentProvider) {
+  if (currentProvider) {
     browser.storage.local.get('providers', (res) => {
-      changes.currentProvider = { newValue: res.providers[changes.currentProvider.newValue] };// eslint-disable-line
+      const newValue = [];
+      const currentProviderValue = currentProvider.newValue;
+      const { length } = currentProviderValue;
+
+      for (let i = 0; i < length; i += 1) {
+        const element = currentProviderValue[i];
+
+        newValue.push(res.providers[element]);
+      }
+      // TODO refactor make copy of changes instead of altering it? {...changes}?
+      changes.currentProvider = { newValue };// eslint-disable-line
       updateLocals(changes);
-      handleMessage(lastMsg);
     });
   } else {
     updateLocals(changes);
-    handleMessage(lastMsg);
+  }
+}
+
+function onPortConnected(port) {
+  if (port.name === 'popup') {
+    port.onDisconnect.addListener(() => {
+      const options = browser.extension.getViews({ type: 'tab' });
+
+      options.forEach(tab => tab.location.reload());
+    });
   }
 }
 
 function init(initSettings) {
   const settings = initSettings || {};
+  const { currentProvider: settingsCurrentProvider } = settings;
+  const currentProvider = [];
+
+  if (typeof settingsCurrentProvider === 'string') { // backward compatibility: after changing type of currenProvider from string to array
+    currentProvider.push(settings.providers[settingsCurrentProvider]);
+  } else if (Array.isArray(settingsCurrentProvider)) {
+    const { length } = settingsCurrentProvider;
+
+    for (let i = 0; i < length; i += 1) {
+      const element = settingsCurrentProvider[i];
+
+      currentProvider.push(settings.providers[element]);
+    }
+  }
 
   // TODO use constant for name
   window.contextMenuSearchLocals = {
-    currentProvider: settings.providers[settings.currentProvider],
+    currentProvider,
     defaultProtocol: settings.defaultProtocol,
-    urlDetected: false,
+    silent: settings.silent,
+    mode: settings.mode,
+    multisearch: settings.multisearch,
+    searchMenu: false, // boolean or array - false if no menu at all; array of ids of context menu items
+    gotoMenu: false,
   };
 
-  const locals = window.contextMenuSearchLocals;
-
-  browser.contextMenus.create({ // TODO Refactor move function outside
-    id: 'contextSearch',
-    title: `${browser.i18n.getMessage('searchWith')} ${locals.currentProvider.name}: "%s"`,
-    contexts: ['selection'],
-    onclick(event) {
-      const query = event.selectionText.trim().replace(/\s/gi, '+');
-
-      browser.tabs.create({
-        url: `${locals.currentProvider.url}${query}`,
-      });
-    },
-  }, () => {
-    browser.runtime.onMessage.addListener(handleMessage);
-    browser.storage.onChanged.addListener(onStorageChange);
-  });
+  browser.runtime.onMessage.addListener(handleMessage);
+  browser.storage.onChanged.addListener(onStorageChange);
+  browser.runtime.onConnect.addListener(onPortConnected);
 }
 
 function setDefaultStoreValues() {
   browser.storage.local.get(null, (res) => {
+    // TODO refactor make copy of res instead of altering it
     const result = res;
     let shouldUpdate;
 
@@ -164,11 +254,23 @@ function setDefaultStoreValues() {
     }
     if (!result.currentProvider) {
       shouldUpdate = true;
-      result.currentProvider = 'google';
+      result.currentProvider = ['google'];
     }
     if (!result.defaultProtocol) {
       shouldUpdate = true;
       result.defaultProtocol = 'https://';
+    }
+    if (!result.silent) {
+      shouldUpdate = true;
+      result.silent = false;
+    }
+    if (!result.mode) {
+      shouldUpdate = true;
+      result.mode = '3';
+    }
+    if (!result.multisearch) {
+      shouldUpdate = true;
+      result.multisearch = false;
     }
 
     if (shouldUpdate) {
@@ -182,4 +284,3 @@ function setDefaultStoreValues() {
 if (!window.browser) window.browser = chrome; // Compatibility for Chrome
 
 setDefaultStoreValues();
-