+++ src/options.js
@@ -1,10 +1,11 @@
 if (!window.browser) window.browser = chrome; // Compatibility for Chrome
 
-function createOption(value, text) {
+function createOption(value, text, selected) {
   const el = document.createElement('option');
 
   el.value = value;
   el.innerText = text;
+  el.selected = selected;
 
   return el;
 }
@@ -17,17 +18,24 @@ function createProviderDiv(name, url) {
   return el;
 }
 
-function initSelect(providers, currentProvider) {
+function initSelect(providers, currentProvider, isMultisearch) {
   const select = document.getElementById('searchprovider');
+  const current = Array.isArray(currentProvider) ? currentProvider : [currentProvider];
 
+  select.multiple = !!isMultisearch;
   Object.keys(providers).forEach((v) => {
-    select.appendChild(createOption(v, providers[v].name));
+    const selected = current.includes(v);
+
+    select.appendChild(createOption(v, providers[v].name, selected));
   });
-  select.value = currentProvider;
+
+  const validity = document.getElementById('providervalidity');
+
+  validity.style.display = select.selectedOptions.length ? 'none' : 'block';
 }
 
 function resetCustomProviders() {
-  browser.storage.local.get('providers', (result) => {
+  browser.storage.local.get(['multisearch', 'providers'], (result) => {
     const providers = {};
     const resProviders = result.providers;
 
@@ -40,32 +48,39 @@ function resetCustomProviders() {
     });
 
     // TODO current provider global(external) constant
-    browser.storage.local.set({ providers, currentProvider: 'google' }, () => {
+    browser.storage.local.set({ providers, currentProvider: ['google'] }, () => {
       document.getElementById('providerslist').innerText = '';
       document.getElementById('searchprovider').innerText = '';
       // TODO current provider global(external) constant
-      initSelect(providers, 'google');
-      browser.contextMenus.update('contextSearch', {
-        title: `${browser.i18n.getMessage('searchWith')} ${providers.google.name}: "%s"`,
-      });
+      initSelect(providers, ['google'], result.multisearch);
     });
   });
 }
 
 function changeProvider() {
-  browser.storage.local.set({ currentProvider: this.value }, () => {
-    browser.storage.local.get('providers', (result) => {
-      const currentProvider = result.providers[this.value];
+  const { options } = this;
+  const { length } = options;
+  const values = [];
 
-      browser.contextMenus.update('contextSearch', {
-        title: `${browser.i18n.getMessage('searchWith')} ${currentProvider.name}: "%s"`,
-      });
-    });
-  });
+  for (let i = 0; i < length; i += 1) {
+    const element = options[i];
+
+    if (element.selected) {
+      values.push(element.value);
+    }
+  }
+
+  const validity = document.getElementById('providervalidity');
+
+  validity.style.display = values.length ? 'none' : 'block';
+
+  browser.storage.local.set({ currentProvider: values });
 }
 
 function loadCustomProvidersList(list, container) {
-  Object.keys(list).forEach((v) => {
+  const listArray = Object.keys(list);
+
+  listArray.forEach((v) => {
     const provider = list[v];
 
     if (provider.custom) {
@@ -104,6 +119,7 @@ function addProvider() {
   }
 }
 
+// TODO refactor replace with universal function setInStore(storageKey, elementKey, [callback])
 function changeProtocol() {
   browser.storage.local.set({ defaultProtocol: this.value });
 }
@@ -115,17 +131,88 @@ function initProtocol(protocol) {
   select.addEventListener('change', changeProtocol);
 }
 
+function changeSilent() {
+  browser.storage.local.set({ silent: this.checked });
+}
+
+function initSilent(isSilent) {
+  const checkbox = document.getElementById('silent');
+
+  checkbox.checked = isSilent;
+  checkbox.addEventListener('change', changeSilent);
+}
+
+function changeMode() {
+  browser.storage.local.set({ mode: this.value });
+}
+
+function initMode(mode) {
+  const group = document.getElementsByName('mode');
+  const { length } = group;
+
+  for (let i = 0; i < length; i += 1) {
+    const element = group[i];
+
+    if (element.value === mode) {
+      element.checked = true;
+    } else {
+      element.checked = false;
+    }
+    element.addEventListener('change', changeMode);
+  }
+}
+
+function changeMultisearch() {
+  const { checked } = this;
+
+  browser.storage.local.set({ multisearch: checked }, () => {
+    const select = document.getElementById('searchprovider');
+    const note = document.getElementById('providernote');
+    const validity = document.getElementById('providervalidity');
+
+    note.style.display = checked ? 'block' : 'none';
+    select.multiple = checked;
+    validity.style.display = select.selectedOptions.length ? 'none' : 'block';
+
+    if (!checked) {
+      const e = new Event('change');
+      select.dispatchEvent(e);
+    }
+  });
+}
+
+function initMultisearch(isMultisearch) {
+  const checkbox = document.getElementById('multisearch');
+
+  checkbox.checked = isMultisearch;
+  checkbox.addEventListener('change', changeMultisearch);
+}
+
+function initNote(isMultisearch) {
+  const note = document.getElementById('providernote');
+
+  note.style.display = isMultisearch ? 'block' : 'none';
+}
+
 document.addEventListener('DOMContentLoaded', () => {
   browser.storage.local.get(null, (result) => {
     const providersListEl = document.getElementById('providerslist');
 
-    initSelect(result.providers, result.currentProvider);
+    initSelect(result.providers, result.currentProvider, result.multisearch);
     loadCustomProvidersList(result.providers, providersListEl);
 
     document.getElementById('searchprovider').addEventListener('change', changeProvider);
     document.getElementById('addbutton').addEventListener('click', addProvider);
     document.getElementById('resetbutton').addEventListener('click', resetCustomProviders);
-
+    // TODO Separate init and addEventListener() ?
     initProtocol(result.defaultProtocol);
+
+    initSilent(result.silent);
+
+    initMode(result.mode);
+
+    initMultisearch(result.multisearch);
+
+    initNote(result.multisearch);
   });
 });